Feature/destination folder: Added support for specifying destination folder.

+++ src/file_utils.rs
@@ -1,4 +1,4 @@
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 
 use walkdir::{DirEntry, WalkDir};
 use zip_extensions::{zip_create_from_directory, zip_extract};
@@ -8,8 +8,8 @@ use crate::{
     path_utils::{self},
 };
 
-pub fn process_files(comic_directory: String) -> () {
-    let comic_files = filter_comic_files(comic_directory);
+pub fn process_files(comic_directory: String, destination_directory: Option<String>) -> () {
+    let comic_files = filter_comic_files(&comic_directory);
 
     // Initialize and configure the progress bar for comics.
     let comic_progress_bar =
@@ -30,9 +30,25 @@ pub fn process_files(comic_directory: String) -> () {
 
         process_images(&temp_dir);
 
-        path_utils::remove_file(&comic_pathbuf, comic_pathbuf.to_str().unwrap());
+        // Create destination file path from comic file path based on destination directory.
+        let dest_path = match destination_directory {
+            Some(ref destination_directory) => {
+                let relative_path = comic_pathbuf.strip_prefix(&comic_directory).unwrap();
+                let dest_path = Path::new(&destination_directory).join(relative_path);
+                if let Some(parent) = dest_path.parent() {
+                    std::fs::create_dir_all(parent).expect("Failed to create destination directories");
+                }
+                dest_path
+            }
+            None => comic_pathbuf.clone(),
+        };
 
-        zip_create_from_directory(&comic_pathbuf, &temp_dir).unwrap();
+        // Remove original comic file if original comic file is not specified.
+        if destination_directory.is_none() {
+            path_utils::remove_file(&comic_pathbuf, comic_pathbuf.to_str().unwrap());
+        }
+
+        zip_create_from_directory(&dest_path, &temp_dir).unwrap();
 
         path_utils::remove_directory(&temp_dir.to_path_buf(), &temp_dir.to_str().unwrap());
     }
@@ -62,7 +78,7 @@ fn process_images(images_directory: &PathBuf) -> () {
 /*
 Walk through the comic directory and create a vector of all the comic files.
 */
-fn filter_comic_files(comic_directory: String) -> Vec<DirEntry> {
+fn filter_comic_files(comic_directory: &String) -> Vec<DirEntry> {
     let comic_files = WalkDir::new(comic_directory)
         .into_iter()
         // Filter the comic files out of the directories and unrelated files
